Rename files and structure them after their collection
