+++ lib/connectRecipes.js
@@ -1,9 +1,5 @@
 const { collection } = require("./database");
-const {
-  deleteOneElement,
-  findOneElement,
-  findAndSort,
-} = require("./helpFunctions");
+const { findOneElement, findAndSort } = require("./helpFunctions");
 
 async function getRecipiesMongo() {
   return findAndSort("recipes", { RecipeName: 1 });
